requirements.txt
pytest
requests
PyYAML
python-dotenv
faker
pytest-html
allure-pytest